pipeline/conversational_agent/pyproject.toml
[project]
name = "conversational-agent"
version = "0.1.0"
description = "SHIFT Conversational Agent Pipeline - GROW coaching model agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "langchain>=1.0",
    "langchain-anthropic",
    "langgraph-checkpoint-firestore",
    "google-cloud-firestore",
    "google-cloud-secret-manager",
    "pydantic>=2.7.4",
    "python-jose[cryptography]==3.3.0",
    "httpx==0.25.2",
    "cachetools==5.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

